Initialize snapshot state before loading tests
I'm not sure under which circumstances this exactly goes wrong, but this
fixes a difference in behaviour that I observed between the upstream
Jest runner and this runner.

A test file that worked with the upstream runner would error out with
the following message:

    Snapshot state must be initialized

    Snapshot state has value: undefined

Initializing the snapshot state before loading the tests fixes this.